Allow proper log output during test runs.
This involved a huge amount of cleanup to existing code
in order to make sure that existing failures were properly
fixed or silenced when intentional.

Additionally, this introduces a set of pre-built exception
instances for test purposes which are marked with the
"silence_logging" attribute to denote expected failures which
should not be logged in test output.

Change-Id: I6b420382dcdc5792c7be4d727853f438154c641e

Quantum Integration with Horizon

- All CRUD operations for networks and ports.
- Cann attach ports to instance interfaces.
- Tests added for API wrapper and views.
- All operations currently in user panels.

Change-Id: I0680a9da2908573c0804c7af0cbafd569ea48573

Fixing the quantum API

Fixing quantum tests for the new quntumclient.